+++ app/api/users.py
@@ -35,7 +35,7 @@ router = APIRouter()
 
 
 @router.get("/", responses=response_schemas.single_users_responses)
-def get_user(user: User = Depends(get_user)) -> JSONResponse:
+def get_user_me(user: User = Depends(get_user)) -> JSONResponse:
     if user is None:
         return JSONResponse(status_code=500,
                             content={"message": "Internal Server Error"})
@@ -59,9 +59,13 @@ def get_user(user: User = Depends(get_user)) -> JSONResponse:
 
 @router.put("/", responses=response_schemas.single_users_responses)
 @manage_role_access(UserRoleEnum.MANAGER)
-def update_user(data: UserUpdate,
-                db: Session = Depends(deps.get_db),
-                user: User = Depends(get_user)) -> JSONResponse:
+def update_user(user_id: int,
+                data: UserUpdate,
+                db: Session = Depends(deps.get_db)) -> JSONResponse:
+    user = crud_users.get_user_by_id(user_id, db)
+    if user is None:
+        return JSONResponse(status_code=500,
+                            content={"message": "Internal Server Error"})
     data = crud_users.update_user(user, update=data, db=db)
     if data is None:
         return JSONResponse(status_code=500,
@@ -73,8 +77,11 @@ def update_user(data: UserUpdate,
 
 @router.delete("/", responses=response_schemas.general_responses)
 @manage_role_access(UserRoleEnum.MANAGER)
-def delete_user(db: Session = Depends(deps.get_db),
-                user: User = Depends(get_user)) -> JSONResponse:
+def delete_user(user_id: int, db: Session = Depends(deps.get_db)) -> JSONResponse:
+    user = crud_users.get_user_by_id(user_id, db)
+    if user is None:
+        return JSONResponse(status_code=500,
+                            content={"message": "Internal Server Error"})
     data = crud_users.delete_user(user, db=db)
     if data is None:
         return JSONResponse(status_code=500,
@@ -107,3 +114,15 @@ def accept_terms_user(db: Session = Depends(deps.get_db),
 
     return JSONResponse(status_code=200,
                         content={"message": "success"})
+
+
+@router.get("/", responses=response_schemas.single_users_responses)
+@manage_role_access(UserRoleEnum.MANAGER)
+def get_user_id(user_id: int, db: Session = Depends(deps.get_db)) -> JSONResponse:
+    user = crud_users.get_user_by_id(user_id, db)
+    if user is None:
+        return JSONResponse(status_code=500,
+                            content={"message": "Internal Server Error"})
+    json_compatible_item_data = jsonable_encoder(user)
+    return JSONResponse(status_code=200,
+                        content=json_compatible_item_data)
